Refractoring the project with MVC but error on editText view, it says that we try to invoke method on null object
