+++ app/src/main/java/Vue/BaseActivity.java
@@ -1,4 +1,4 @@
-package Controler;
+package Vue;
 
 import android.content.Context;
 import android.view.View;
