+++ app/src/main/res/layout/log_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Model.LogActivity">
+    tools:context="Vue.LogActivity">
 
     <ProgressBar
         android:id="@+id/progressBar"
